third_party/WebKit/LayoutTests/editing/deleting/delete_br.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>one</div>',
            '|<br>',
            '<div>two</div>',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            '<div>one|</div>',
            '<div>two</div>',
        '</div>',
    ].join('')),
    '1 Delete BR after caret');

test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>one</div>',
            '<br>',
            '<div>|two</div>',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            '<div>one</div>',
            '<div>|two</div>',
        '</div>',
    ].join('')),
    '2 Delete BR before caret');

test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div>one</div>',
            '<div>|two</div>',
        '</div>',
    ].join(''),
    'delete',
    [
        '<div contenteditable>',
            '<div>one|two</div>',
        '</div>',
    ].join('')),
    '3 Delete across DIVs');
</script>
